+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - added `detail` field to `Participant`, passed in event payloads, `user_disconnected`, `connection_closed`, `connection_error`, or undefined
+- added `msElapsed` field to the `onTimer` handler
 
 ## [0.4.0] - 2022-02-11